+++ src/server/types.ts
@@ -1,4 +1,13 @@
-import type { CliFilteringOptions } from "../toolFilters.ts";
+import { z } from "zod";
+import { CliFilteringOptionsSchema } from "../toolFilters.ts";
 
+export const StartServerOptionsSchema = CliFilteringOptionsSchema.extend({
+  credentials: z
+    .object({
+      applicationId: z.string(),
+      apiKey: z.string(),
+    })
+    .optional(),
+});
 
-export type StartServerOptions = CliFilteringOptions;
+export type StartServerOptions = z.infer<typeof StartServerOptionsSchema>;
